# openapi/application-service.yaml
openapi: 3.0.3
info:
  title: GACP Application Service API
  description: |
    GACP certification application management service with 12-state workflow.

    **Features:**
    - Application CRUD operations
    - 12-state finite state machine
    - Document upload and validation
    - DTAM review workflow
    - State transition management
    - Timeline estimation

    **Application States:**
    1. DRAFT - Initial state
    2. SUBMITTED - Farmer submitted
    3. UNDER_REVIEW - DTAM reviewing
    4. PAYMENT_PENDING - Awaiting Phase 1 payment
    5. PAYMENT_VERIFIED - Payment confirmed
    6. INSPECTION_SCHEDULED - Inspection date set
    7. INSPECTION_COMPLETED - Inspection done
    8. PHASE2_PAYMENT_PENDING - Awaiting Phase 2 payment
    9. PHASE2_PAYMENT_VERIFIED - Phase 2 paid
    10. APPROVED - Application approved
    11. CERTIFICATE_ISSUED - Certificate generated
    12. REJECTED - Application rejected
    13. REVISION_REQUIRED - Needs revision
    14. EXPIRED - Payment timeout

    **Research Base:**
    - Thai FDA certification workflow
    - WHO GMP certification standards
    - Finite State Machine patterns (GoF Design Patterns)
  version: 1.0.0
  contact:
    name: GACP Platform API Support
    email: api-support@gacp.platform
servers:
  - url: https://api.gacp.platform/v1/applications
    description: Production server
  - url: https://staging-api.gacp.platform/v1/applications
    description: Staging server
  - url: http://localhost:3002/v1/applications
...
paths:
...
  /applications/{applicationId}/documents:
    post:
      tags:
        - Documents
      summary: Upload document
      description: |
        Upload document to application.

        **File Requirements:**
        - Formats: PDF, JPG, PNG
        - Max size: 10MB per file
        - Required types: ID_CARD, HOUSE_REGISTRATION, LAND_OWNERSHIP

        **Authorization:** Application owner only
      operationId: uploadDocument
      security:
        - BearerAuth: []
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
                - documentType
              properties:
                file:
                  type: string
                  format: binary
                  description: Document file (PDF, JPG, PNG)
                documentType:
                  type: string
                  enum:
                    - ID_CARD
                    - HOUSE_REGISTRATION
                    - LAND_OWNERSHIP
                    - OTHER
                  description: Document type
      responses:
        '201':
          description: Document uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentUploadResponse'
        '400':
          description: Invalid file or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '413':
          description: File too large (max 10MB)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DocumentUploadResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Document uploaded successfully
        data:
          type: object
          properties:
            documentId:
              type: string
            type:
              type: string
            fileName:
              type: string
            fileSize:
              type: integer
            uploadedAt:
              type: string
              format: date-time
            url:
              type: string
              format: uri
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            field:
              type: string
            details:
              type: object
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...